refactor: simplify weaviate user cross references reset
