+++ _schemaV2.graphql
@@ -16890,7 +16890,7 @@ type Query {
     limit: Int
     offset: Int
     sort: DiscoverArtworksSort
-    userId: String
+    userId: String!
   ): ArtworkConnection
 
   # A namespace external partners (provided by Galaxy)
@@ -21567,7 +21567,7 @@ type Viewer {
     limit: Int
     offset: Int
     sort: DiscoverArtworksSort
-    userId: String
+    userId: String!
   ): ArtworkConnection
 
   # A namespace external partners (provided by Galaxy)
